fix(sim): faithful T7 offense/defense from the betterSkillys stat formulas
A fidelity audit of Player/StatsManager showed the sim modeled the fight far harder than the real
game on offense and mis-modeled the gear stats. Corrected against source, with every value verified:

OFFENSE (was ~3x too weak):
- Staff attack multiplier: player projectile dmg = raw * GetAttackMult, mult = 0.5 + ATT/75*1.5;
  maxed Wizard ATT 75 -> x2.0. Staff of Destruction raw 45-85 -> 90-170 (baked into staff_dmg). The
  BulletNova spell does NOT scale (AEBulletNova uses raw Random.Next) -> spell_dmg stays raw 95-185.
- Staff fire rate: GetAttackFreq at DEX 75 = 0.008 -> 125ms/shot = 1.25 ticks, not 2. staff_cooldown
  is now a FRACTIONAL accumulator (carry the sub-tick remainder, snap idle back to 0) for ~8 shots/s.
- Staff range: staff_life 6 -> 4.75 ticks (Life 475ms).

GEAR STATS (the equip "stat" id is a StatDataType resolved via GetStatIndex, NOT a raw core-stat id;
the earlier read of stat 3 as Defense was wrong -- stat 3 = MaxMana, Defense = stat 21):
- player_defense 30 -> 8 (robe stat21=Defense 8; base Wizard DEF 0). The squishier-but-correct value.
- player_hp_max 770 -> 810 (670 base + 100 ring stat0 + 40 spell stat0).
- player_mp_max 385 -> 455 (385 base + 30 robe stat3=MaxMana + 40 spell stat3=MaxMana).
- mp_regen 1.54 -> 1.78 (WIS 70 = 60 base + 3 robe + 7 spell; (1 + 0.24*WIS)/s).
- player_speed 0.55 -> 0.773 (MoveSpeed at SPD 50: 0.004 + 50/75*(0.0096-0.004) t/ms).

All four Config surfaces synced; test_config_sync + numpy-oracle<->C parity green. test_parity_boss_fight
boss_hp bumped 20000 -> 54000 so the higher T7 DPS doesn't kill it inside the 150-step parity window
(the boss died identically in both sims at step 142 -- a stale window constant, not a parity break).

Deferred (noted for later, lower impact): wavy staff shots (Amplitude/Frequency), BulletNova
emitting at the cursor/boss rather than the player, ScaleHP2(20) solo boss-HP scaling (kept 7500 to
match the eval/video tooling), petrify diagonal-fan angles, two folded snake types.

+++ puffer4/binding.c
@@ -38,12 +38,12 @@ void my_init(Env* env, Dict* kwargs) {
     c->damage_floor = dict_get(kwargs, "damage_floor")->value;
     c->mp_regen = dict_get(kwargs, "mp_regen")->value;
     c->hp_regen = dict_get(kwargs, "hp_regen")->value;
-    c->staff_cooldown = (int)dict_get(kwargs, "staff_cooldown")->value;
+    c->staff_cooldown = dict_get(kwargs, "staff_cooldown")->value;
     c->staff_num = (int)dict_get(kwargs, "staff_num")->value;
     c->staff_dmg_lo = dict_get(kwargs, "staff_dmg_lo")->value;
     c->staff_dmg_hi = dict_get(kwargs, "staff_dmg_hi")->value;
     c->staff_speed = dict_get(kwargs, "staff_speed")->value;
-    c->staff_life = (int)dict_get(kwargs, "staff_life")->value;
+    c->staff_life = dict_get(kwargs, "staff_life")->value;
     c->staff_radius = dict_get(kwargs, "staff_radius")->value;
     c->staff_offset = dict_get(kwargs, "staff_offset")->value;
     c->spell_cost = dict_get(kwargs, "spell_cost")->value;
